Just go with a pip requirements file
Walked back on the new setup.py install structure and added a requirements.txt file for installation of dependencies via pip. Added some lines to the beginning of the example "run_" script that allow bxh2bids to be imported even if the "run_" script is not in the package directory.